Updated purge code to better handle aborted backups
The purge code will normally look for the oldest in-progress backup,
and not purge any files older than that date.  This is a problem if a
backup was aborted or crashed, as it looks like an in-progress backup.
This fix will ignore any such backup logs older than 3 days.
(This should become a tunable number.)